# openapi/okoscope-v1.yaml
openapi: 3.1.0
info:
  title: Okoscope Web API
  version: 1.0.0
  description: Stable API contract for the separately deployed Okoscope Web UI. Timestamps are RFC 3339 UTC values.
servers:
  - url: /
security:
  - sessionAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RegisterRequest' } } }
      responses:
        '201': { $ref: '#/components/responses/AuthResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/auth/login:
    post:
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/AuthResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      responses:
        '200': { $ref: '#/components/responses/AuthResponse' }
        '401': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/auth/logout:
    post:
      operationId: logoutUser
      responses:
        '204': { description: Session revoked and cookie expired }
        '403': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: okoscope_session
      description: Opaque HttpOnly browser session; clients send requests with credentials enabled.
    adminAuth: { type: http, scheme: bearer, description: System administrator credential }
...
  headers:
    RequestId: { schema: { type: string }, description: Correlation identifier for the request. }
  schemas:
    OrganizationRole: { type: string, enum: [owner, member] }
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, organization_slug, organization_name]
      properties:
        email: { type: string, format: email, minLength: 3, maxLength: 254 }
        password: { type: string, format: password, minLength: 12, maxLength: 256, writeOnly: true }
        organization_slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 63 }
        organization_name: { type: string, minLength: 1, maxLength: 120 }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, minLength: 3, maxLength: 254 }
        password: { type: string, format: password, minLength: 1, maxLength: 256, writeOnly: true }
    AuthenticatedUser:
      type: object
      additionalProperties: false
      required: [id, email]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        email: { type: string, format: email }
    AuthenticatedOrganization:
      type: object
      additionalProperties: false
      required: [id, slug, name]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        slug: { type: string }
        name: { type: string }
    AuthContext:
      type: object
      additionalProperties: false
      required: [user, organization, role]
      properties:
        user: { $ref: '#/components/schemas/AuthenticatedUser' }
        organization: { $ref: '#/components/schemas/AuthenticatedOrganization' }
        role: { $ref: '#/components/schemas/OrganizationRole' }
...
    Error:
      type: object
      additionalProperties: false
      required: [error, message, request_id]
      properties:
        error: { type: string, example: not_found }
        message: { type: string, example: resource not found }
        request_id: { type: string, example: 0ec02ed2-8483-4981-893e-bffc535897d7 }
        fields:
          type: object
          additionalProperties: { type: string }
          description: Present for field-level validation failures.
...
    Uuid: { type: string, format: uuid }
...
  responses:
    AuthResponse:
      description: Safe authenticated user, Organization, and role context; the opaque session is set only as an HttpOnly cookie.
      headers:
        X-Request-Id: { schema: { type: string } }
        Set-Cookie: { schema: { type: string, writeOnly: true } }
        Cache-Control: { schema: { type: string, const: no-store } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AuthContext' }
...
    Error:
      description: Correlated API error
      headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }